+++ src/main/java/bc1/gream/domain/gifticon/repository/GifticonRepository.java
@@ -1,6 +1,6 @@
 package bc1.gream.domain.gifticon.repository;
 
-import bc1.gream.domain.order.entity.Gifticon;
+import bc1.gream.domain.gifticon.entity.Gifticon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GifticonRepository extends JpaRepository<Gifticon, Long>, GifticonRepositoryCustom {
