+++ src/test/java/bc1/gream/domain/sell/repository/SellRepositoryCustomImplTest.java
@@ -4,7 +4,7 @@ import static bc1.gream.domain.product.entity.QProduct.product;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import bc1.gream.domain.order.entity.Gifticon;
+import bc1.gream.domain.gifticon.entity.Gifticon;
 import bc1.gream.domain.product.dto.response.SellPriceToQuantityResponseDto;
 import bc1.gream.domain.sell.entity.QSell;
 import bc1.gream.domain.sell.entity.Sell;
