+++ src/test/java/bc1/gream/domain/sell/repository/SellRepositoryTest.java
@@ -2,8 +2,8 @@ package bc1.gream.domain.sell.repository;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import bc1.gream.domain.gifticon.entity.Gifticon;
 import bc1.gream.domain.gifticon.repository.GifticonRepository;
-import bc1.gream.domain.order.entity.Gifticon;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.repository.ProductRepository;
 import bc1.gream.domain.sell.entity.Sell;
